tratamento de erro NotFound na Categoria

+++ src/main/java/com/example/cupcat/service/CategoriaService.java
@@ -2,22 +2,20 @@ package com.example.cupcat.service;
 
 import com.example.cupcat.dto.CategoriaDTO;
 import com.example.cupcat.exception.AlreadyExistingException;
-import com.example.cupcat.exception.NotFoundException;
 import com.example.cupcat.model.Categoria;
-import com.example.cupcat.model.Produto;
 import com.example.cupcat.repository.CategoriaRepo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
-import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
 public class CategoriaService implements ICategoria{
     private final CategoriaRepo repo;
+    private static final String MSG_ERROR_NOT_FOUND = "Categoria não encontrada!";
 
     @Override
     public void save(CategoriaDTO categoriaDTO) throws AlreadyExistingException {
@@ -31,18 +29,29 @@ public class CategoriaService implements ICategoria{
     }
 
     @Override
-    public Optional<Categoria> getCategoriaById(int id) throws NotFoundException {
+    public Optional<Categoria> getCategoriaById(int id) throws NoSuchElementException {
+        Optional<Categoria> opCat = repo.findById(id);
+
+        if(opCat.isEmpty()) throw new NoSuchElementException(MSG_ERROR_NOT_FOUND);
         return repo.findById(id);
     }
 
     @Override
-    public void updateCategoria(Categoria categoria) throws NotFoundException {
+    public void updateCategoria(CategoriaDTO categoriaDTO, int id) throws NoSuchElementException {
+        if(!repo.existsById(id)) throw new NoSuchElementException(MSG_ERROR_NOT_FOUND);
+
+        Categoria categoria = new Categoria(categoriaDTO);
+        categoria.setId(id);
+
         repo.save(categoria);
     }
 
     @Override
-    public Optional<Categoria> removeCategoriaById(int id) throws NotFoundException {
+    public Optional<Categoria> removeCategoriaById(int id) throws NoSuchElementException {
         Optional<Categoria> categoria = this.getCategoriaById(id);
+
+        if(categoria.isEmpty()) throw new NoSuchElementException(MSG_ERROR_NOT_FOUND);
+
         repo.deleteById(id);
 
         return categoria;
